chore: move config command into configuration.ts
