+++ packages/codegen/src/commands.ts
@@ -6,57 +6,9 @@ import {
   PackageOption,
   ServiceOption,
   closestConfigDirectoryPath,
-  VersionOption,
 } from './utils';
 import { CodeGen, passThroughFilter, pathFilterFromConfig } from './codegen';
 import { ConfigManager, PackageConfig, ServiceConfig } from './configuration';
-
-interface CreateConfigOptions
-  extends PackageOption,
-    Partial<ServiceOption>,
-    Partial<VersionOption> {}
-
-export class ConfigureCommand {
-  public create = async (options: CreateConfigOptions) => {
-    const directory = await closestConfigDirectoryPath(options);
-    const mngr = new ConfigManager({
-      directory: directory,
-    });
-    if (await mngr.exists()) {
-      throw new Error(`Config already exists - did nothing`);
-    }
-
-    const config = this.defaultConfig(options);
-    await mngr.write(config);
-  };
-
-  private defaultConfig = (
-    options: PackageOption & Partial<ServiceOption> & Partial<VersionOption>
-  ): ServiceConfig | PackageConfig => {
-    if (options.service != null) {
-      return {
-        service: options.service,
-        filter: {
-          serviceName: options.service,
-        },
-        inlinedSchemas: {
-          autoNameRequest: true,
-        },
-      } as ServiceConfig;
-    } else {
-      if (options.version == null) {
-        throw new Error(
-          '"Version" must be defined when creating a package config'
-        );
-      }
-      return {
-        snapshot: {
-          version: options.version,
-        },
-      } as PackageConfig;
-    }
-  };
-}
 
 type GenerateOptions = PackageOption;
